Removed all previous attached event handlers from elements
Every time a new search is performed, a new event handler
is attached to the elements. This affects the toggle behavior
of the modal. Sometimes clicking the `click for more...` text at
the bottom of the teaser text doesn't open the modal.

The modal not opening is due to another event handler attached
to the elements. Even though the previous search items are replaced
with new search items, the previous event handlers are not removed.

Previous wrong behavior, case 1:
1. Search for title.
2. Clicking text `click for more...` doesn't open modal.
3. Refresh page.
4. Search for the SAME title.
5. Clicking text `click for more...` now opens the the modal.

Previous wrong behavior, case 2:
1. Search for title.
2. One-third of the time, clicking `click for more...` will
open the modal.
3. Other two-thirds of the time, behavior is same as case 1.

Previous wrong behavior, case 3:
1. The modal will work on the first search term.
2. The modal will NOT work on the second search term.
3. Observed, the modal doesn't seem to work on ODD numbered searches
but works on EVEN numbered searches.